META from enigma (MetaSupport not needed anymore)
